pyproject.toml
[project]
name = "plugin-video-netflix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "codecov>=2.1.13",
    "kodi-addon-checker>=0.0.36",
    "kodistubs==19.0.3",
    "mysql-connector-python>=9.4.0",
    "polib>=1.2.0",
    "pycryptodomex>=3.23.0",
    "pylint>=3.3.8",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "tox>=4.29.0",
    "xmlschema>=4.1.0",
]
